Revise Schnorr components
