+++ src/libCrypto/Schnorr.cpp
@@ -39,542 +39,24 @@
 
 using namespace std;
 
-std::mutex BIGNUMSerialize::m_mutexBIGNUM;
-std::mutex ECPOINTSerialize::m_mutexECPOINT;
-
 Curve::Curve()
     : m_group(EC_GROUP_new_by_curve_name(NID_secp256k1), EC_GROUP_clear_free),
       m_order(BN_new(), BN_clear_free) {
-  if (m_order == nullptr) {
-    LOG_GENERAL(WARNING, "Curve order setup failed");
-    // throw exception();
-  }
-
   if (m_group == nullptr) {
-    LOG_GENERAL(WARNING, "Curve group setup failed");
-    // throw exception();
+    LOG_GENERAL(FATAL, "Curve group setup failed");
+    return;
   }
-
+  if (m_order == nullptr) {
+    LOG_GENERAL(FATAL, "Curve order setup failed");
+    return;
+  }
   // Get group order
   if (!EC_GROUP_get_order(m_group.get(), m_order.get(), NULL)) {
-    LOG_GENERAL(WARNING, "Recover curve order failed");
-    // throw exception();
+    LOG_GENERAL(FATAL, "Recover curve order failed");
   }
 }
 
 Curve::~Curve() {}
-
-shared_ptr<BIGNUM> BIGNUMSerialize::GetNumber(const bytes& src,
-                                              unsigned int offset,
-                                              unsigned int size) {
-  if (size <= 0) {
-    LOG_GENERAL(WARNING, "assertion failed (" << __FILE__ << ":" << __LINE__
-                                              << ": " << __FUNCTION__ << ")");
-    return nullptr;
-  }
-
-  // This mutex is to prevent multi-threaded issues with the use of openssl
-  // functions
-  lock_guard<mutex> g(m_mutexBIGNUM);
-
-  if (offset + size <= src.size()) {
-    BIGNUM* ret = BN_bin2bn(src.data() + offset, size, NULL);
-    if (ret != NULL) {
-      return shared_ptr<BIGNUM>(ret, BN_clear_free);
-    }
-  } else {
-    LOG_GENERAL(WARNING, "Unable to get BIGNUM of size "
-                             << size << " from stream with available size "
-                             << src.size() - offset);
-  }
-
-  return nullptr;
-}
-
-void BIGNUMSerialize::SetNumber(bytes& dst, unsigned int offset,
-                                unsigned int size, shared_ptr<BIGNUM> value) {
-  if (size <= 0) {
-    LOG_GENERAL(WARNING, "assertion failed (" << __FILE__ << ":" << __LINE__
-                                              << ": " << __FUNCTION__ << ")");
-    return;
-  }
-
-  // This mutex is to prevent multi-threaded issues with the use of openssl
-  // functions
-  lock_guard<mutex> g(m_mutexBIGNUM);
-
-  const int actual_bn_size = BN_num_bytes(value.get());
-
-  // if (actual_bn_size > 0)
-  {
-    if (actual_bn_size <= static_cast<int>(size)) {
-      if (offset + size > dst.size()) {
-        dst.resize(offset + size);
-      }
-
-      // Pad with zeroes as needed
-      const unsigned int size_diff =
-          size - static_cast<unsigned int>(actual_bn_size);
-      fill(dst.begin() + offset, dst.begin() + offset + size_diff, 0x00);
-
-      if (BN_bn2bin(value.get(), dst.data() + offset + size_diff) !=
-          actual_bn_size) {
-        LOG_GENERAL(WARNING, "Unexpected serialized size");
-      }
-    } else {
-      LOG_GENERAL(WARNING, "BIGNUM size ("
-                               << actual_bn_size
-                               << ") exceeds requested serialize size (" << size
-                               << ")");
-    }
-  }
-  // else
-  // {
-  //     LOG_MESSAGE("Error: Zero-sized BIGNUM");
-  // }
-}
-
-shared_ptr<EC_POINT> ECPOINTSerialize::GetNumber(const bytes& src,
-                                                 unsigned int offset,
-                                                 unsigned int size) {
-  shared_ptr<BIGNUM> bnvalue = BIGNUMSerialize::GetNumber(src, offset, size);
-
-  // This mutex is to prevent multi-threaded issues with the use of openssl
-  // functions
-  lock_guard<mutex> g(m_mutexECPOINT);
-
-  if (bnvalue != nullptr) {
-    unique_ptr<BN_CTX, void (*)(BN_CTX*)> ctx(BN_CTX_new(), BN_CTX_free);
-    if (ctx == nullptr) {
-      LOG_GENERAL(WARNING, "Memory allocation failure");
-      // throw exception();
-      return nullptr;
-    }
-
-    EC_POINT* ret =
-        EC_POINT_bn2point(Schnorr::GetInstance().GetCurve().m_group.get(),
-                          bnvalue.get(), NULL, ctx.get());
-    if (ret != NULL) {
-      return shared_ptr<EC_POINT>(ret, EC_POINT_clear_free);
-    }
-  }
-  return nullptr;
-}
-
-void ECPOINTSerialize::SetNumber(bytes& dst, unsigned int offset,
-                                 unsigned int size,
-                                 shared_ptr<EC_POINT> value) {
-  shared_ptr<BIGNUM> bnvalue;
-  {
-    // This mutex is to prevent multi-threaded issues with the use of openssl
-    // functions
-    std::lock_guard<mutex> g(m_mutexECPOINT);
-
-    unique_ptr<BN_CTX, void (*)(BN_CTX*)> ctx(BN_CTX_new(), BN_CTX_free);
-    if (ctx == nullptr) {
-      LOG_GENERAL(WARNING, "Memory allocation failure");
-      // throw exception();
-      return;
-    }
-
-    bnvalue.reset(
-        EC_POINT_point2bn(Schnorr::GetInstance().GetCurve().m_group.get(),
-                          value.get(), POINT_CONVERSION_COMPRESSED, NULL,
-                          ctx.get()),
-        BN_clear_free);
-    if (bnvalue == nullptr) {
-      LOG_GENERAL(WARNING, "Memory allocation failure");
-      // throw exception();
-      return;
-    }
-  }
-
-  BIGNUMSerialize::SetNumber(dst, offset, size, bnvalue);
-}
-
-PrivKey::PrivKey() : m_d(BN_new(), BN_clear_free), m_initialized(false) {
-  // kpriv->d should be in [1,...,order-1]
-
-  if (m_d != nullptr) {
-    const Curve& curve = Schnorr::GetInstance().GetCurve();
-
-    m_initialized = true;
-    do {
-      if (!BN_rand_range(m_d.get(), curve.m_order.get())) {
-        LOG_GENERAL(WARNING, "Private key generation failed");
-        m_initialized = false;
-        break;
-      }
-    } while (BN_is_zero(m_d.get()));
-  } else {
-    LOG_GENERAL(WARNING, "Memory allocation failure");
-    // throw exception();
-  }
-}
-
-PrivKey::PrivKey(const bytes& src, unsigned int offset) {
-  if (Deserialize(src, offset) != 0) {
-    LOG_GENERAL(WARNING, "We failed to init PrivKey.");
-  }
-}
-
-PrivKey::PrivKey(const PrivKey& src)
-    : m_d(BN_new(), BN_clear_free), m_initialized(false) {
-  if (m_d != nullptr) {
-    if (BN_copy(m_d.get(), src.m_d.get()) == NULL) {
-      LOG_GENERAL(WARNING, "PrivKey copy failed");
-    } else {
-      m_initialized = true;
-    }
-  } else {
-    LOG_GENERAL(WARNING, "Memory allocation failure");
-    // throw exception();
-  }
-}
-
-PrivKey::~PrivKey() {}
-
-PrivKey PrivKey::GetPrivKeyFromString(const string& key) {
-  if (key.size() != 64) {
-    throw std::invalid_argument(
-        "Error: private key - invalid number of input characters for key");
-  }
-  bytes key_v;
-
-  try {
-    if (!DataConversion::HexStrToUint8Vec(key, key_v)) {
-      throw std::invalid_argument(
-          "Error: public  key - invalid number of input characters for key");
-    }
-  } catch (std::exception& e) {
-    throw std::invalid_argument(
-        "Error: private key - invalid format of input characters for key - "
-        "required hexadecimal characters");
-  }
-
-  return PrivKey(key_v, 0);
-}
-
-bool PrivKey::Initialized() const { return m_initialized; }
-
-unsigned int PrivKey::Serialize(bytes& dst, unsigned int offset) const {
-  // LOG_MARKER();
-
-  if (m_initialized) {
-    BIGNUMSerialize::SetNumber(dst, offset, PRIV_KEY_SIZE, m_d);
-  }
-
-  return PRIV_KEY_SIZE;
-}
-
-int PrivKey::Deserialize(const bytes& src, unsigned int offset) {
-  // LOG_MARKER();
-
-  try {
-    m_d = BIGNUMSerialize::GetNumber(src, offset, PRIV_KEY_SIZE);
-    if (m_d == nullptr) {
-      LOG_GENERAL(WARNING, "Deserialization failure");
-      m_initialized = false;
-      return -1;
-    } else {
-      m_initialized = true;
-    }
-  } catch (const std::exception& e) {
-    LOG_GENERAL(WARNING, "Error with PrivKey::Deserialize." << ' ' << e.what());
-    return -1;
-  }
-  return 0;
-}
-
-PrivKey& PrivKey::operator=(const PrivKey& src) {
-  m_initialized = (BN_copy(m_d.get(), src.m_d.get()) == m_d.get());
-  return *this;
-}
-
-bool PrivKey::operator==(const PrivKey& r) const {
-  return (m_initialized && r.m_initialized &&
-          (BN_cmp(m_d.get(), r.m_d.get()) == 0));
-}
-
-PubKey::PubKey()
-    : m_P(EC_POINT_new(Schnorr::GetInstance().GetCurve().m_group.get()),
-          EC_POINT_clear_free),
-      m_initialized(false) {
-  if (m_P == nullptr) {
-    LOG_GENERAL(WARNING, "Memory allocation failure");
-    // throw exception();
-  }
-}
-
-PubKey::PubKey(const PrivKey& privkey)
-    : m_P(EC_POINT_new(Schnorr::GetInstance().GetCurve().m_group.get()),
-          EC_POINT_clear_free),
-      m_initialized(false) {
-  if (m_P == nullptr) {
-    LOG_GENERAL(WARNING, "Memory allocation failure");
-    // throw exception();
-    return;
-  } else if (!privkey.Initialized()) {
-    LOG_GENERAL(WARNING, "Private key is not initialized");
-    return;
-  } else {
-    const Curve& curve = Schnorr::GetInstance().GetCurve();
-
-    if (BN_is_zero(privkey.m_d.get()) ||
-        (BN_cmp(privkey.m_d.get(), curve.m_order.get()) != -1)) {
-      LOG_GENERAL(WARNING,
-                  "Input private key is invalid. Public key "
-                  "generation failed");
-      return;
-    }
-    if (EC_POINT_mul(curve.m_group.get(), m_P.get(), privkey.m_d.get(), NULL,
-                     NULL, NULL) == 0) {
-      LOG_GENERAL(WARNING, "Public key generation failed");
-      return;
-    }
-
-    m_initialized = true;
-  }
-}
-
-PubKey::PubKey(const bytes& src, unsigned int offset) {
-  if (Deserialize(src, offset) != 0) {
-    LOG_GENERAL(WARNING, "We failed to init PubKey.");
-  }
-}
-
-PubKey::PubKey(const PubKey& src)
-    : m_P(EC_POINT_new(Schnorr::GetInstance().GetCurve().m_group.get()),
-          EC_POINT_clear_free),
-      m_initialized(false) {
-  if (m_P == nullptr) {
-    LOG_GENERAL(WARNING, "Memory allocation failure");
-    // throw exception();
-    return;
-  } else if (src.m_P == nullptr) {
-    LOG_GENERAL(WARNING, "src (ec point) is null in pub key construct.");
-    // throw exception();
-  } else {
-    if (EC_POINT_copy(m_P.get(), src.m_P.get()) != 1) {
-      LOG_GENERAL(WARNING, "PubKey copy failed");
-    } else {
-      m_initialized = true;
-    }
-  }
-}
-
-PubKey::~PubKey() {}
-
-bool PubKey::Initialized() const { return m_initialized; }
-
-unsigned int PubKey::Serialize(bytes& dst, unsigned int offset) const {
-  if (m_initialized) {
-    ECPOINTSerialize::SetNumber(dst, offset, PUB_KEY_SIZE, m_P);
-  }
-
-  return PUB_KEY_SIZE;
-}
-
-PubKey PubKey::GetPubKeyFromString(const string& key) {
-  if (key.size() != 66) {
-    throw std::invalid_argument(
-        "Error: public  key - invalid number of input characters for key");
-  }
-  bytes key_v;
-
-  try {
-    if (!DataConversion::HexStrToUint8Vec(key, key_v)) {
-      throw std::invalid_argument(
-          "Error: public  key - invalid number of input characters for key");
-    }
-  } catch (std::exception& e) {
-    throw std::invalid_argument(
-        "Error: public key - invalid format of input characters for key - "
-        "required hexadecimal characters");
-  }
-
-  return PubKey(key_v, 0);
-}
-
-int PubKey::Deserialize(const bytes& src, unsigned int offset) {
-  // LOG_MARKER();
-
-  try {
-    m_P = ECPOINTSerialize::GetNumber(src, offset, PUB_KEY_SIZE);
-    if (m_P == nullptr) {
-      LOG_GENERAL(WARNING, "Deserialization failure");
-      m_initialized = false;
-      return -1;
-    } else {
-      m_initialized = true;
-    }
-  } catch (const std::exception& e) {
-    LOG_GENERAL(WARNING, "Error with PubKey::Deserialize." << ' ' << e.what());
-    return -1;
-  }
-  return 0;
-}
-
-PubKey& PubKey::operator=(const PubKey& src) {
-  m_initialized =
-      src.m_initialized && (EC_POINT_copy(m_P.get(), src.m_P.get()) == 1);
-  return *this;
-}
-
-bool PubKey::operator<(const PubKey& r) const {
-  unique_ptr<BN_CTX, void (*)(BN_CTX*)> ctx(BN_CTX_new(), BN_CTX_free);
-  if (ctx == nullptr) {
-    LOG_GENERAL(WARNING, "Memory allocation failure");
-    // throw exception();
-    return false;
-  }
-
-  shared_ptr<BIGNUM> lhs_bnvalue(
-      EC_POINT_point2bn(Schnorr::GetInstance().GetCurve().m_group.get(),
-                        m_P.get(), POINT_CONVERSION_COMPRESSED, NULL,
-                        ctx.get()),
-      BN_clear_free);
-  shared_ptr<BIGNUM> rhs_bnvalue(
-      EC_POINT_point2bn(Schnorr::GetInstance().GetCurve().m_group.get(),
-                        r.m_P.get(), POINT_CONVERSION_COMPRESSED, NULL,
-                        ctx.get()),
-      BN_clear_free);
-
-  return (m_initialized && r.m_initialized &&
-          (BN_cmp(lhs_bnvalue.get(), rhs_bnvalue.get()) == -1));
-}
-
-bool PubKey::operator>(const PubKey& r) const {
-  unique_ptr<BN_CTX, void (*)(BN_CTX*)> ctx(BN_CTX_new(), BN_CTX_free);
-  if (ctx == nullptr) {
-    LOG_GENERAL(WARNING, "Memory allocation failure");
-    // throw exception();
-    return false;
-  }
-
-  shared_ptr<BIGNUM> lhs_bnvalue(
-      EC_POINT_point2bn(Schnorr::GetInstance().GetCurve().m_group.get(),
-                        m_P.get(), POINT_CONVERSION_COMPRESSED, NULL,
-                        ctx.get()),
-      BN_clear_free);
-  shared_ptr<BIGNUM> rhs_bnvalue(
-      EC_POINT_point2bn(Schnorr::GetInstance().GetCurve().m_group.get(),
-                        r.m_P.get(), POINT_CONVERSION_COMPRESSED, NULL,
-                        ctx.get()),
-      BN_clear_free);
-
-  return (m_initialized && r.m_initialized &&
-          (BN_cmp(lhs_bnvalue.get(), rhs_bnvalue.get()) == 1));
-}
-
-bool PubKey::operator==(const PubKey& r) const {
-  unique_ptr<BN_CTX, void (*)(BN_CTX*)> ctx(BN_CTX_new(), BN_CTX_free);
-  if (ctx == nullptr) {
-    LOG_GENERAL(WARNING, "Memory allocation failure");
-    // throw exception();
-    return false;
-  }
-
-  return (m_initialized && r.m_initialized &&
-          (EC_POINT_cmp(Schnorr::GetInstance().GetCurve().m_group.get(),
-                        m_P.get(), r.m_P.get(), ctx.get()) == 0));
-}
-
-Signature::Signature()
-    : m_r(BN_new(), BN_clear_free),
-      m_s(BN_new(), BN_clear_free),
-      m_initialized(false) {
-  if ((m_r == nullptr) || (m_s == nullptr)) {
-    LOG_GENERAL(WARNING, "Memory allocation failure");
-    // throw exception();
-  } else {
-    m_initialized = true;
-  }
-}
-
-Signature::Signature(const bytes& src, unsigned int offset) {
-  if (Deserialize(src, offset) != 0) {
-    LOG_GENERAL(WARNING, "We failed to init Signature.");
-  }
-}
-
-Signature::Signature(const Signature& src)
-    : m_r(BN_new(), BN_clear_free),
-      m_s(BN_new(), BN_clear_free),
-      m_initialized(false) {
-  if ((m_r != nullptr) && (m_s != nullptr)) {
-    m_initialized = true;
-
-    if (BN_copy(m_r.get(), src.m_r.get()) == NULL) {
-      LOG_GENERAL(WARNING, "Signature challenge copy failed");
-      m_initialized = false;
-    }
-
-    if (BN_copy(m_s.get(), src.m_s.get()) == NULL) {
-      LOG_GENERAL(WARNING, "Signature response copy failed");
-      m_initialized = false;
-    }
-  } else {
-    LOG_GENERAL(WARNING, "Memory allocation failure");
-    // throw exception();
-  }
-}
-
-Signature::~Signature() {}
-
-bool Signature::Initialized() const { return m_initialized; }
-
-unsigned int Signature::Serialize(bytes& dst, unsigned int offset) const {
-  // LOG_MARKER();
-
-  if (m_initialized) {
-    BIGNUMSerialize::SetNumber(dst, offset, SIGNATURE_CHALLENGE_SIZE, m_r);
-    BIGNUMSerialize::SetNumber(dst, offset + SIGNATURE_CHALLENGE_SIZE,
-                               SIGNATURE_RESPONSE_SIZE, m_s);
-  }
-
-  return SIGNATURE_CHALLENGE_SIZE + SIGNATURE_RESPONSE_SIZE;
-}
-
-int Signature::Deserialize(const bytes& src, unsigned int offset) {
-  // LOG_MARKER();
-
-  try {
-    m_r = BIGNUMSerialize::GetNumber(src, offset, SIGNATURE_CHALLENGE_SIZE);
-    if (m_r == nullptr) {
-      LOG_GENERAL(WARNING, "Deserialization failure");
-      m_initialized = false;
-      return -1;
-    } else {
-      m_s = BIGNUMSerialize::GetNumber(src, offset + SIGNATURE_CHALLENGE_SIZE,
-                                       SIGNATURE_RESPONSE_SIZE);
-      if (m_s == nullptr) {
-        LOG_GENERAL(WARNING, "Deserialization failure");
-        m_initialized = false;
-        return -1;
-      } else {
-        m_initialized = true;
-      }
-    }
-  } catch (const std::exception& e) {
-    LOG_GENERAL(WARNING,
-                "Error with Signature::Deserialize." << ' ' << e.what());
-    return -1;
-  }
-  return 0;
-}
-
-Signature& Signature::operator=(const Signature& src) {
-  m_initialized = ((BN_copy(m_r.get(), src.m_r.get()) == m_r.get()) &&
-                   (BN_copy(m_s.get(), src.m_s.get()) == m_s.get()));
-  return *this;
-}
-
-bool Signature::operator==(const Signature& r) const {
-  return (m_initialized && r.m_initialized &&
-          ((BN_cmp(m_r.get(), r.m_r.get()) == 0) &&
-           (BN_cmp(m_s.get(), r.m_s.get()) == 0)));
-}
 
 Schnorr::Schnorr() {}
 
@@ -623,21 +105,6 @@ bool Schnorr::Sign(const bytes& message, unsigned int offset, unsigned int size,
 
   if (message.size() < (offset + size)) {
     LOG_GENERAL(WARNING, "Offset and size beyond message size");
-    return false;
-  }
-
-  if (!privkey.Initialized()) {
-    LOG_GENERAL(WARNING, "Private key not initialized");
-    return false;
-  }
-
-  if (!pubkey.Initialized()) {
-    LOG_GENERAL(WARNING, "Public key not initialized");
-    return false;
-  }
-
-  if (!result.Initialized()) {
-    LOG_GENERAL(WARNING, "Signature not initialized");
     return false;
   }
 
@@ -798,16 +265,6 @@ bool Schnorr::Verify(const bytes& message, unsigned int offset,
 
   if (message.size() < (offset + size)) {
     LOG_GENERAL(WARNING, "Offset and size beyond message size");
-    return false;
-  }
-
-  if (!pubkey.Initialized()) {
-    LOG_GENERAL(WARNING, "Public key not initialized");
-    return false;
-  }
-
-  if (!toverify.Initialized()) {
-    LOG_GENERAL(WARNING, "Signature not initialized");
     return false;
   }
